+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Object properties: fixed transform settings sync with transform controls
 - Object properties: fixed visibility settings sync
+- Raycast: filter selectable objects during intersection
 
 ## [0.19.0] - 2026-04-25